Revert "fix rubocop warnings"
This reverts commit 8a8333783423c8e920761c46da0356daa65c1b05.

+++ scripts/elements/platform.rb
@@ -1,5 +1,3 @@
-# frozen_string_literal: true
-
 class Platform
   def initialize(reinstall, rebuild)
     @dist = "#{Paths::PLATFORM}/dist"
@@ -11,7 +9,9 @@ class Platform
     @changes_to_files = ChangeChecker.has_changes?(Paths::PLATFORM, @targets)
   end
 
-  attr_reader :changes_to_files
+  def changes_to_files
+    @changes_to_files
+  end
 
   def clean
     @targets.each do |path|
@@ -46,7 +46,7 @@ class Platform
         Shell.sh 'yarn run build'
         Reporter.done(self, 'build', '')
       end
-    rescue StandardError
+    rescue
       Reporter.failed(self, 'build', '')
       clean
       build
@@ -59,15 +59,13 @@ class Platform
     platform_dest = "#{node_modules}/platform"
     platform = Platform.new(false, false)
     Dir.mkdir(node_modules) unless File.exist?(node_modules)
-    if replace || !File.exist?("#{platform_dest}/dist") || File.symlink?(platform_dest) || platform.changes_to_files
-      Shell.rm_rf(platform_dest)
+    Shell.rm_rf(platform_dest) if replace || !File.exist?("#{platform_dest}/dist") || File.symlink?(platform_dest) || platform.changes_to_files
+    unless File.exist?(platform_dest)
+      Reporter.other('Platform', "#{consumer} doesn't have platform", '')
+      platform.build
+      Shell.sh "cp -r #{Paths::PLATFORM} #{node_modules}"
+      Reporter.done('Platform', "delivery to #{consumer}", '')
     end
-    return if File.exist?(platform_dest)
-
-    Reporter.other('Platform', "#{consumer} doesn't have platform", '')
-    platform.build
-    Shell.sh "cp -r #{Paths::PLATFORM} #{node_modules}"
-    Reporter.done('Platform', "delivery to #{consumer}", '')
   end
 
   def lint
